refactor: 클린 + 레이어드 아키텍처 구성
