--- src/main/java/com/hanghae/clean_architecture/CapacityStatus.java
@@ -1,5 +1,0 @@
-package com.hanghae.clean_architecture;
-
-public enum CapacityStatus {
-    FULL, EMPTY
-}
