--- src/main/java/com/hanghae/clean_architecture/LectureService.java
@@ -1,8 +1,0 @@
-package com.hanghae.clean_architecture;
-
-import java.util.List;
-
-public interface LectureService {
-
-    List<LectureResponse> searchLectures(LectureSearch lectureSearch);
-}
