--- src/main/java/com/hanghae/clean_architecture/ReservationStatus.java
@@ -1,5 +1,0 @@
-package com.hanghae.clean_architecture;
-
-public enum ReservationStatus {
-    ACTIVE, INACTIVE
-}
