+++ src/main/java/com/hanghae/clean_architecture/domain/lecture/Lecture.java
@@ -1,5 +1,6 @@
-package com.hanghae.clean_architecture;
+package com.hanghae.clean_architecture.domain.lecture;
 
+import com.hanghae.clean_architecture.interfaces.response.lecture.LectureResponse;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
