+++ src/main/java/com/hanghae/clean_architecture/domain/lecture/constant/CapacityStatus.java
@@ -1,0 +1,5 @@
+package com.hanghae.clean_architecture.domain.lecture.constant;
+
+public enum CapacityStatus {
+    FULL, EMPTY
+}
