+++ src/main/java/com/hanghae/clean_architecture/domain/reservation/constant/ReservationStatus.java
@@ -1,0 +1,5 @@
+package com.hanghae.clean_architecture.domain.reservation.constant;
+
+public enum ReservationStatus {
+    ACTIVE, INACTIVE
+}
