+++ src/main/java/com/hanghae/clean_architecture/interfaces/request/lecture/LectureSearch.java
@@ -1,4 +1,4 @@
-package com.hanghae.clean_architecture;
+package com.hanghae.clean_architecture.interfaces.request.lecture;
 
 import lombok.Getter;
