+++ src/main/java/com/hanghae/clean_architecture/interfaces/response/lecture/LectureResponse.java
@@ -1,4 +1,4 @@
-package com.hanghae.clean_architecture;
+package com.hanghae.clean_architecture.interfaces.response.lecture;
 
 import lombok.Getter;
